+++ NEWS.md
@@ -1,3 +1,21 @@
+# WRFmet 0.1.7
+
+* Changed layer_vectorField() to take in a single raster object `raster` and the
+names of its u and v layers: `uName` and `vName`.
+* Changed layer_raster() to take in an optional `varName` parameter in case the
+given raster is a multilayer *RasterBrick*.
+* Changed the names of `plot_base()`, `plot_raster()`, and `plot_standard()` to
+`wrf_basePlot()`, `wrf_rasterPlot()`, and `wrf_standardPlot()` respectively.
+* `wrf_rasterPlot()` and `wrf_standardPlot()` plot titles default to their
+raster title.
+* Updated all vignettes and functions, and examples to properly use these new 
+parameters and function names.
+
+# WRFmet 0.1.6
+
+* `wrf_load()` now sets the title of the loaded *RasterBrick* to the format:
+"[model domain] yyyy-mm-dd hh[am/pm], Hour [forecast hour]"
+
 # WRFmet 0.1.5
 
 * Changed default polygon fill from 'white' to 'transparent' for `layer_states()`